began condensing LF plots, realized LF need change
I started by trying to condense all of the leapfrog plots a bit, specifically by creating another data table that had many o fthe calculations already performed, by using the dev.new.lf() that I'd made in a previous commit, etc. However, I realized that it'd be easier to just make some of the changes in leapfrog to begin with, which led to a reorganization and general improvement of that script.