complete logic for trajectory adjustment
Trajectories are easy to compute when the starting and ending locations are both in the sea. However, additional steps need to be taken when the proposed location is on land. I've written several functions to find the warmest or coolest neighbors (rook search), test which neighbor the trajectory should prefer (pos traj go to cool neighbors, opposite for neg), convert focal cell numbers to global cell numbers, and output the appropriate destination cell number given the information about starting and proposed temperatures and their locations and starting velocity (for sign). It is the destination cell that is needed, which is the proposed cell adjusted for coastline effects etc. These functions are likely completely or close to complete, but have no yet been tested. Time for lunch.