working (?), except prob w/ raster res
The functions I wrote to do the search for the local (rook) minimum/maximum temperature values, and to convert those located cells numbers to global cell numbers, were written assuming the original resolution of the data. So there are many cases where the entire rook search is just the same temperature value. Thus, even when the proposed cell location is failing, it can't finder a higher or lower temperature in area b/c it's all same temperature. The result is that many cases are being classified as coastal sinks. For some reason there's also an issue where there are a lot of NA's in areas that should be classified as sinks/cool/warm. Hmm.