fix?2: interp fills too much; now corrected
When doing the bilinear interpolation, there were cells that were becoming non-NA that were NA when doing the "constant" interpolation. E.g., the surface area of the great lakes apparently increases, cuba disappears, baja peninsula (what was left of it) disappears, etc etc.

To solve that problem, convert all non-NA elements in the "constant" interpolated map (what I was originally using but was giving me problems b/c not finding cooler/ warmer neighbors) to 1, and leave all NA as-is. Then take this raster of 1's and NA's, and multiply it by the "bilinear"-ly interpolated raster. This reconverts cuba, baja peninsula, and land immediately surrounding the great lakes (e.g.) back to NA's (as they should be – land is NA).