first run of trajectories complete
some small issues may remain, but this should be close to final trajectory algorithm. The time steps and spatial resolution may be a bit coarse, especially compared to what burrows did. Run time was a bit long, too. Maybe 15 minutes or so, I didn't time it. A lot of the calculation time is spent finding the local minima and maxima. It may be possible to only perform this search for the part of the map for which it is needed, but this would require a substantial rework, possibly (I've felt a bit constrained by the raster package). Regardless, I still need to profile the run to know for sure what needs optimizing. But that'll come later. Much later.


Former-commit-id: 774630719e420efdce8feb4784ca782255d3fc66