reorg traj prep; adding correct rook vel/angl
Did a bunch of reorganizing for the order of defining variables prior to the iterative determination of trajectories. The reorganization was necessary to pre-calculate the rook velocities and trajectories (w/o making a huge mess of things), but also reduced the number of objects that needed to be defined.

My current thinking is that instead of doing the belWarm*cellWarm (e.g.), I will do something that's more to the effect of belRook4 <- belAdj&(focalCell==4L); belRook4*dYkm.rook

I.e., I am not going to use the logic*raster line to get cell numbers, but instead to get the proper change in latitude or longitude (or maybe I'll do it in km?). Adjustments may need to be made outside of adjDest(), as well. It depends on whether or not it makes more sense to grab all cell #'s inide adjDest() (and perhaps other factors)


Former-commit-id: 7293241ac407b1c4159834922bcb5da985230750