change limitV so can be used for all velocities
limitV was initially created when limiting the rook velocities, and only needed the starting latitudes. While looping through time to create the trajectories, a cell in the raster actually represents the starting location of a trajectory, not the current location of a trajectory. Thus, cell 1,1 is only guarunteed to have the lon-lat of 1,1 at the first time step, afterwards need to reference another raster that keeps track of the location of 1,1. Kind of confusing because initially you could put temperature at 1,1, and know the temperature for that place on the map based on the extent() of the raster. However, after the first time step, a temperature at 1,1 could instead be the temperature at y,x, because that's the current location of the trajectory that started at 1,1.

So just making the conversion factor generic so as to accept the appropriate latitude for each grid cell (b/c grid cell location doesn't necessarily reflect geographic location past 1st time step).


Former-commit-id: 9625af6a9eb2478a261867c92ca6d538fffb7f6b [formerly e0237e9b4e0547244305dd86005ba666273ab3bd]
Former-commit-id: 0511a7e8511bab6e775eb7b74c0d4339908fe3ad