address problematic NA's in spatial gradient
this was a giant pain. I took a tierd approach to solving the problem. For slopes I would do a larger and larger spatial average of the slopes calculated from the 3x3 disaggregated base-slopes (slopes calculated from the 1x1 sst grid). When I needed to fill in the NA's of the base-slopes, I would use the smallest spatial average that still had finite values. There were only 2 spatial grid sizes: 3x3 and 7x7. However, for the 7x7, I also calculated the spatial slope (to be 7x7 averaged) from a bilinearly interpolated sst. So that was another level of abstraction occurring in parallel.

I tried to take a similar approach to the angles of the spatial slopes. However, I problem here is that you can't average 0 degrees and 360 degrees and get the right direction. So I had to do spatial averaging of the temperatures, then calculate the angles from there. The lack of consistency to how I approach the slope and the angle might need to be addressed. However, I thought it better to average the slope than to average the temperature, that the more complicated calculation is tied back to the less/un transformed data.

Also with the angle, the saf7.vals isn't needed, so I'll remove that in a subsequent commit


Former-commit-id: a4df209dc093450a21a38bbdccef417c6a816836