dealing with spatial angles
I *do* need the f7 angles, because before there were edges where none of the spatial averages were givings values (but where there were sst's); didn't notice this was a problem because values of 0 were being returned (oops)

Also, there's a lot of commented-out code that I've added. I tried to do the angles just like a did the slopes – by doing a spatial average of the high-level calculation (slope/ angle), rather than a spatial average of sst then calculating the slope/ angle on the averaged sst. However, I was 1) getting a scale between -3 and 3, not between 0 and 3, so I didn't know what to do with that; 2) even still the scale bar didn't seem the same, so I don't know if it was working properly. Taking the spatial average of sst temperature seemed to produce patterns that were more similar to the unaveraged slopes. One problem with the f7 angle is that there are pretty clear boundary issues, but f7 is only needed at the edge of the maps, and the f7 boundary issues are in the middle of the ocean usually. So might be fine. Leaving commented code for now, just in case.


Former-commit-id: 58f19dc30ff009543a63bd03e8dd69dfdad1486b