cleaning adjDest; error missing cool/warmLon/Lat
There are places where startLon or startLat are not NA, but cool- warmLon/Lat are NA. Obviously this would happen over land. Actually, now that I'm looking at graphs more carefullly, I think this is only happening over land (and maybe the great lakes as well). So maybe this is a problem that coul dbe fixed by replacing places with missing startVel with the startLon or startLat

Good news is that this isn't a problem with the resolution or mismatch between spatial - temporal gradients and the sst's or anything like that. (I just fixed that issue!)


Former-commit-id: 00a43114f83577e9926ee07e8d87b663fe871497 [formerly 661014579a7f8ef572e33151ff927287e9ffb2c7]
Former-commit-id: 5779d7f1588ec35906b0200e4e9378065796a2ab