fix problem with invalid rooked trajectories
solution: just set the new destination to the center of the rook, rather than moving trajectory along in the destination of the ... wait a minute. crap. i may have made this more complicated than it had to be. There shouldn't be both Y and X components to the rook trajectories. Because the trajectory is only supposed to move along in the direction of the rook.

Former-commit-id: d963800eccb13a8607fd13be3c1123b9c842b997 [formerly 17627437439abf460eb7d734c816b1c5dc452b3d]
Former-commit-id: def5aec109f44a3f22291fb74b6d9388df09ed18