fix problem with invalid rooked trajectories
solution: just set the new destination to the center of the rook, rather than moving trajectory along in the destination of the ... wait a minute. crap. i may have made this more complicated than it had to be. There shouldn't be both Y and X components to the rook trajectories. Because the trajectory is only supposed to move along in the direction of the rook.

Former-commit-id: 17627437439abf460eb7d734c816b1c5dc452b3d