remove negative rook velocities?
It was this change that got me thinking that there was a mixup in the calculations of rook velocities and directions with respect to 0 vs 90 north conventions. I might have to undo this change .. or at least i don't think it should be necessary ... or maybe just take the absolute value? Gosh, confusing.


Former-commit-id: c7bbda633bb5355fbc1b6db1904e9f5378c991a3