focal.max now pads with -Inf values instead of NA
I think that this should help with correctly identifying the best focal cell ... I was getting a lot more weird values when I wasn't doing this. I think part of the problem i the encoding of the NA and NaN values in R (which is as a number, maybe platform specific), and how that converts into and out of a raster object. Because I was getting a lot of values that were x*10^314, e.g.


Former-commit-id: f12164c89018a6a3288ddae7cc5a1b625dcc69b6