function to convert raster NaN's to NA's
It was seeming that NaN's were possibly being interpreted as odd values. Coincidentally (and more relevant to the true motivation for this function), the 6 bad cool Rooks I was getting were NaN's at the destination. So I just wanted to be sure that it wasn't a NaN NA issue. It wasn't.  But nothing wrong with being consistent here.

O yeah, also, it was only values around the coast that were NaN. Nothing on the interior. Very odd.


Former-commit-id: 34c394e04480fa43a1721115809a6b98afdfe785 [formerly a360607f658f20c189a1108ee7005264e2431f41]
Former-commit-id: 6dee4258b4ae2be48945c7c91074b4e5826cd5ca