Big: Complete re-haul of adjusting bad proposals
This is a pretty big set of changes. After a couple days of fiddling with my initial strategy for adjusting trajectories, I decided to give up on it. I could not figure out why a very small number of the adjusted trajectories were invalid. And the formatting of the data in raster objects made it extremely hard to explore. Also, computationally intensive b/c all possible adjustments were made for all possible trajectories. In the end, I could get the "right" adjusted destinations, it seemed, but I had to force it into the center of the rook cell.

One important change I made that could have been causing some of those problems was that I had originally set the sequence of lons and lats as starting and ending at the extrema of the map. However, it makes more sense to have the trajectory starting locations be at the center of the cell, not its edges. This was even causing duplicates (from the first time step!) when doing cellFromXY()!

My new approach was inspired by my discovery of adjacent(). This function returns a matrix of cell numbers for neighbors. My previous approach had been based on wrapping which.min() in focal(). This was both computationally intensive (b/c which.min is applied to *every* cell) and hard to work with (b/c the returned object was HUGE [see previous] and it was a raster).

Adjustments are now made on an as-needed basis. Calculations are still made for all 5 rook possibilities, but only for the focal cells that need adjusting. This is an immense change in computation time.

This method remains untested. However, a promising characteristic of the approach is that it is quite easy to get all relevant information about starting, proposed, and adjusted locations: cell#, lon, lat, temperature, angle, velocity, etc. It's easier b/c there are less attempts to adjust and it's in a matrix format. Much easier to visual than many extensive rasters.


Former-commit-id: 9be862ac28eecd7a0fe5931d8a3132babd620fd5