first shot at categories
I still need to aggregate() the trajectory endings, otherwise convergence is impossible b/c there is only 1 starting trajectory per unaggregated grid, to have convergence you need to have fewer trajectories ending than starting in a location, but the number of trajectories ending there has to be >0. In other words, you need 1>x, but x!=0 (x is the number of ending trajectories, and thus must be >=0). So that's why the Convergence plot is all 0 (FALSE).


Former-commit-id: e5e9eec0477de93986c5d1299ae91607776fa0a9