putting sGulf on 1º grid
I ran into a problem here: If I only used strata (1º definition) that were observed every year, I would only have 5 strata (this was 12 strata w/ original definition). If I allow 1 year to be missing, this goes up to about 16 strata. The increase is pretty linear w/ tolerance – the # of strata you keep is ~ proportional to the number of years you allow to be missing.

I decided to allow 1 year to be missing.


Former-commit-id: 9aea1263e2087b04fe8ce657d258548307623f25 [formerly a6e475011913a2b142a6999d1bccf4108e0312d2]
Former-commit-id: 8be7ce508e0a6e25836a1cd3f6fbab86de874b90