putting sGulf on 1º grid
I ran into a problem here: If I only used strata (1º definition) that were observed every year, I would only have 5 strata (this was 12 strata w/ original definition). If I allow 1 year to be missing, this goes up to about 16 strata. The increase is pretty linear w/ tolerance – the # of strata you keep is ~ proportional to the number of years you allow to be missing.

I decided to allow 1 year to be missing.


Former-commit-id: 722cc9b17b13869608b6ab47b28e1995539974c7 [formerly 99893faf9cc2ac476180dfed9e043dfca44ac143]
Former-commit-id: 6f3e6a9d8d5811c73ab47c6255b175db8bcc8856