putting sGulf on 1º grid
I ran into a problem here: If I only used strata (1º definition) that were observed every year, I would only have 5 strata (this was 12 strata w/ original definition). If I allow 1 year to be missing, this goes up to about 16 strata. The increase is pretty linear w/ tolerance – the # of strata you keep is ~ proportional to the number of years you allow to be missing.

I decided to allow 1 year to be missing.


Former-commit-id: 99893faf9cc2ac476180dfed9e043dfca44ac143