refine "trawl", now time series padded w/ NA's etc
It took me almost a full day to make these changes.

Much of this reflect the changes that were to be made in diverseData. Decided to do it here b/c I wanted the trawl object to be more polished and formatted.

Will need to change diverseData so that it doesn't try to repeat any of these steps.

Furthermore, I think that I may have improved upon some of the corrections made while originally creating diverseData


Former-commit-id: eec976a3edb81cdbe9598c32908e0af37dba8532 [formerly af862b4f5fc0eaf2cfb2e69618e50a5160b79373]
Former-commit-id: 94a3bb48c3e7407d4e30e1a8c2298e5fbbc1e8ed