re-run diversity for completed 1º; add checks to divData
hadn't done neus, newf, or goa. added checks in divData to make sure that s.reg==wc only had species found in both wctri and wcann

added checks in alpha and beta to make sure that there's wasn't weirdness w/ NA's or 0's. I think this problem initially arose b/c I had forgotten to re-run combine.trawl, and changes I'd made there and in read.xx.R and in divData should have taken care of problem. Specifically, what I noticed was that in calculating alpha, WC and 0's for all alpha and alpha now. This was because some species had only 0's and NA's, and all NA's were in the first part of the time series. So I think that this was because I had padded in combine.trawl. Either way, it's fixed now.


Former-commit-id: de2f4b8317343c3dbf62466247396b92a38ae709