added K to trawl2; fixed spp name bug
the bug occurred near line 226, and was related to combining the taxonomic information that was collected automatically and the information gathered manually. While trying to prevent duplicates when combining these sources, I accidentally excluded species that were present in both. This was 237 "spp" (not necessarily identified to spp). I expect this to significantly augment the size of the data set. Was noticed b/c of Clupea harengus (Atlantic herring) missing. Only had rows with the misspelling "Clupea haregus".


Former-commit-id: dfe6c8ef3011b75f46689b28e27cca87256f6328