added K to trawl2; fixed spp name bug
the bug occurred near line 226, and was related to combining the taxonomic information that was collected automatically and the information gathered manually. While trying to prevent duplicates when combining these sources, I accidentally excluded species that were present in both. This was 237 "spp" (not necessarily identified to spp). I expect this to significantly augment the size of the data set. Was noticed b/c of Clupea harengus (Atlantic herring) missing. Only had rows with the misspelling "Clupea haregus".


Former-commit-id: 4e50d0f9eca4191138e875031f5cb5d4e092aca9 [formerly 89c0dce81aa0bdbaff31d3fa3603fdda1abf3887 [formerly dfe6c8ef3011b75f46689b28e27cca87256f6328]]
Former-commit-id: b36052d2d347e7b820008e9364f09f9a36bf1657
Former-commit-id: 7017ac51ffe758319fdb58f3b6679a2630fdc0b3