disorganized intermediate of combine.trawl update
I am only committing this b/c I want to take a snapshot of some of the approaches I was playing around with, and if I have trouble undoing some sort of overwrite (i.e., saving over trawl on accident), I want to see how it happened.

Probably unnecessary, but just in case

there's a lot of experimental code here. a lot


Former-commit-id: a21b13623bdbd0431ba8e29ea9d06d9730e615b7 [formerly c5b1c083df0dda3b748066b1f64ed5f480531e51 [formerly f7095baa6d8deec82ee61bc544fb511c9901c56c]]
Former-commit-id: ea8f48c5d8230a8b42f52a891a973c0a87c271b9
Former-commit-id: 463bddd315be98bcf42045ec2e24b5d1c9f1f717