disorganized intermediate of combine.trawl update
I am only committing this b/c I want to take a snapshot of some of the approaches I was playing around with, and if I have trouble undoing some sort of overwrite (i.e., saving over trawl on accident), I want to see how it happened.

Probably unnecessary, but just in case

there's a lot of experimental code here. a lot


Former-commit-id: f7095baa6d8deec82ee61bc544fb511c9901c56c