implement expand.data in combine.trawl
huge changes here. tried to streamline the combining and expansion due to adding in missing species-year-stratum combinations, and similar for K (reps), by using the new expand.data function.

The function expand.data() is still new and a bit rough – it was pretty hard to create. Seems stable now, so I think that this will be a lasting approach to manipulating the trawl and msom data in combine.trawl.

it is arguable that some of this should not be done in combine.trawl, as it extends well beyond the bare-bones combining and formatting of regional trawl surveys.

however, for the time being, this seems to be a workable approach.


Former-commit-id: e9343b7e385758728f2bccedd131e7d28cd45c65 [formerly e313f79d3e5e424d020c951bf5972ded178efb23 [formerly dec790fbca302f669eba771e95a64f9083fde565]]
Former-commit-id: 06a9c346cdda9ee553c2b1dc7ba61a3c61f731cd
Former-commit-id: 4f8bbbeb56971b714f780de428a1289734b5dcf7