plot categories & trajectories, shelf and not; #15
But for some reason both the shelf and not-shelf restrictions produce almost identical outcomes.

It looks like the trajectories are probably crossing the shelf boundary.

However, it's not obvious (to me) why my code is allowing this – it looks like I've explicitly forbidden this type of movement

Inspecting the old (non-shelf) trajectories shows that in the last iteration (b/c this is what is saved from the image), "toTemp" was NA hundreds of times or something crazy like that. What ever is causing (allowing) that could very well be the cause of #13.


Former-commit-id: f2c9db9c8d8701027ac462cdad3dac5ae56d46f9 [formerly 56a348914cce395e989917f5d709ca7751d7a969]
Former-commit-id: cfac5f5d9ae4d7149255ba1968858d6f68edea0d