simplify covariate formatting, move to format script
I was previously doing a lot of the covariate formatting in the richness.cov script, but really it should have been in format.msom.cov.

Also, I am no longer calculating the precisions from the aggregated mean. If a mean (across K and species) was actually calculated, it was given a very large (1E4) precision. If there wasn't a temperature, e.g., observed in a stratum for a any of the species, then I used the grand mean to fill in, and the precision was based on the sd of the existing per-stratum means.


Former-commit-id: 610bdb1bf04c5e6c921bce37837bdd1de357f307