handle shelf 2011, aggregate spp time, combine WC
the scotian shelf in 2011 didn't have any bottom temperature data, so those parameters had to be omitted from the model. Need to handle this when combining output, else error.

Create objects that aggregate over spp and time in rco – many of the RCO columns were constant for a region, year, or species, leading to many duplicates. These aggregations are convenient for later calculations.

Combine the west coast triennial and annual surveys by dropping annual in the overlapping year (2003), and renaming the region to "WC"