massive post-Rachel update to trawl taxonomy
did a *lot* of checking

I think we have 2557 taxa. The increase in # of taxa is largely due to a lot of my checking/ definition changes, and partly due to newly confirmed taxa via Rachel. Really hard to explain, point is that Rachel helped a lot, but that a lot of the differences are due to extra careful checking and defining on my part (i.e., could have been done before, possibly ... I think, not sure)


Former-commit-id: 73efa03d3b16e525516bb7576b95e19fdc5c4446