wrote colonize function to populate a grid
takes a vector of temperatures, the density of species across temperatures, and the mean biomass of species. returns probability of inhabiting based on temperature, p/a for each cell, and biomass for each cell. Based off code previously in script. Dressed up a little and made clearer and a bit more concise.

works for many or 1 species