use rlnorm() to generate rbeta() params
This is done so that the distribution of the sd() of each species' thermal occupancy is more skewed. The idea is that I want most species to have a narrow thermal occupancy, and thus most species to be rare ... but a few have a broad thermal niche and therefore be common. This setup doesn't directly dictate the species distribution ... it's several steps removed. But I think it's the reason that the MSOM was previously overestimating richness (species had similar niche size, and all niches are observed ~equally).


Former-commit-id: ac4c4e1845d874a0b4d4af94b1ced722ae23bea9