use relative=TRUE in dsample; makes more sense
makes more sense from the standpoint of how a logistic regression would work. The idea here is that in the best thermal habitat, the adjustment from the baseline probability of being present is a factor of 1; i.e., no change. If we do a logistic regression on this result, it works better than when relative=F, because the "density" depends on the discrete nature of the x-axis. It would makse sense to *nnot* rescale it if we wanted to use it as a PDF and calculate the probability that an observation originated from thsi distribution.