fix & demonstrate proper scaling of dnorm for psi
helps #87

You don't want to divide by the max for dnorm, because that'll mean that in some of the observed values there's always one that's going to be perfect. dsample can divide by max because it takes a range of hypothetical x values that we know cover the full range of x values in the simulation. It is with this "full universe" of x values that the scaling of y's is done, and it is only after this step that the y values associated with observed x values are returned. dividing by max dnorm the way I did is tantamount to scaling dsample after the y's are returned, which guarantees that there will always be an x location that has a y of 1. Rather than defining a universe of x's for dnorm (like was done with dsample), I instead just divide (scale) by the density (dnorm) at the quantile equal to the true mu (this will always be max theoretical density). Much more direct. This explanation seems detailed, but it's really simple (only providing long note b/c it wasn't initially apparent to me that I'd need to do it differently, and the proper comparison between probabilities & dnorm and & dsample is something that will need to be done accurately and clearly for a paper)


Former-commit-id: bd166332c16851e5bca532fe894d048e9aff7642