dbl checked calcs; prep'ing for report generation
I had no idea it'd be so easy to do the knitr spin() rmarkdown::render()! Very nice to not have to make a separate document/ script to do this. Just a little extra formatting characters within this script turns it into a nice report. Very cool, much like what I've been looking for.


Former-commit-id: 6c45dec64a82007a31e8b61fe9b6d95963c79c1f