correct Nsite.obs.mu
This may have been accurate previously, but it wasn't taking advantage of more recent work I'd done on the spp attributes in the observation simulation function. I haven't triple checked this yet, but it seems to make sense, and I wasn't tired when I programmed this (well, wasn't feeling too tired, anyway), so it's likely correct.