output Z in right format
argh. I had seen spp.pres[r2m,,] <-  and thought "oh, spp.pres is a matrix now" ... no, it's raster, which is why it had to use the r2m (raster 2 matrix) to receive the G values (which are in matrix format) in the right order.

On a smaller note, returning S.dens.X again. Used in some plots. Probably not the right format anymore b/c it's not a density class, but o well. at least it's there.