increase substrata
this is something I haven't really tried much yet, and might help. I've been noticing that the msom doesn't do much better than observations for site-specific richness, so long as the chance of observing species is decent. But this happens even when the MSOM does OK at estimating psi. It just seems overly fragile to detectability. Currently, the best idea I have for overcoming this fragility is to increase the number of substrata - this should give the model a better handle of detectability via sample size.


Former-commit-id: 8cd623e964dd24927389fbb81a0f5c99014d4a37