reduce latitudinal gradient in temperature
The idea behind this is to increase the density of points at any given temperature. Rather than getting more contrast in the response curve, I think I need to get more density at the critical points. In addition, the changes to the response curves made in 88aca6de and b4888c10 raised the average value of psi just about everywhere by ~3x. This will also help to effectively increase the density of species observations at a given temperature because you'll get more data on the species when they're observed more. You would need a large number of presences and absences for each species to get the curve right.


Former-commit-id: 1c56ab2e8de7dd95d4007368a6ca98f2be8941dc