update Stan to better reflect new model/ data form
Worked a lot today on how to formulate this model. In particular, getting the dimensions of the parameters, data, and covariates organized appropriately for the model, and for Stan. Some of this progress was reflected in the 2 previous commits which was an R file for preparing the data and covariates.