testing with full EBS
#99

This is an interesting test because it'll help reveal how the timing scales with the size of the data set. It is easy to say that timing scales linearly with the number of chain iterations. But how the timing scales with data set dimensions will depend on what dimensions are vectorized over, and how much each dimension grows. This out come will be very telling.

In start of run Chain 1 got to 20% (10 iterations) in ~3 minutes. But other 3 chains were at about 5 iterations, and are hanging. I don't know why some iterations are taking so long.