improve approach to n0's
previously, look at richness in each year, found the year with the most species, and added 50 to that. This number was the target number of species, so n0's in each year was the the number of extra (n0) species needed to reach the target.

now, I look at the total richness across all years, and add 100 to it. Why? Well, richness across all years is at least as high as richness in any given year. So now we're ensuring that in any year the number of species the model is evaluating is greater than the grand total diversity observed. This change, along with doubling the n0_pad, is more conservative with respect to ensuring that the data augmentation is large enough so that each year reaches a similar point along the asymptote.