use info_yrs in bt, but make sure matches rd_yr
there was a problem merging bt and depth into the new bt when rd contained fewer years than the msom output -- it looked like certain strata were duplicated in a given year. To have the merge be more robust I'm now using info_yrs, because info_yrs, bt, and depth are all derived from the msom output (info and inputData), so it makes sense to use the 'year' value from there. At the same time, this shouldn't make a difference because rd (derived from data_all, which is ultimately input to run_msom) should cover the same years as the msom output. So I added a stopifnot() check to make sure that the years agree (which they should, because a while ago I added code to subset rd to years covered in the msom, and I just now added code to subset msom to years covered in rd)